Move EndOfChain class to internal package
